+++ NEWS.md
@@ -1,9 +1,11 @@
-## rtables 0.6.6.9005
+## rtables 0.6.6.9008
 ### New Features
  * Added `top_level_section_div` for `basic_table` to set section dividers for top level rows.
+ * Added `keep_label_rows` to `as_result_df` to have these lines visible.
 
 ### Bug Fixes
  * Fixed `rlistings` decoration (e.g. titles and footers) expansion when there are new lines. Moved relevant handling from `rtables`' `matrix_form` function to `formatters`' dedicated `mform_handle_newlines` function.
+ * Fixed issue with `rtables_root` not being removed when using `as_result_df`.
 
 ## rtables 0.6.6
 ### New Features